Fix tests to account for moving sources widget out
